AI deck working with new CPX restructure
