+++ src/deck/drivers/src/Kbuild
@@ -1,7 +1,6 @@
 obj-y += test/
 obj-$(CONFIG_DECK_ACTIVE_MARKER)    += activeMarkerDeck.o
 obj-$(CONFIG_DECK_AI)               += aideck.o
-obj-$(CONFIG_DECK_AI)               += aideck-router.o
 obj-$(CONFIG_DECK_BIGQUAD)          += bigquad.o
 obj-$(CONFIG_DECK_BUZZ)             += buzzdeck.o
 obj-$(CONFIG_DECK_CPPM)             += cppmdeck.o
